fix(cnr client): to temporarily activate auto-idn conversion for parameter PARENTDOMAIN
We see a backend system issue related to QueryNameserverList where the parameter PARENTDOMAIN is not
automatically idn converted in the CNR Backend System how they do with parameter DOMAIN. By that the
command fails in context of IDNs. Our fix is addressing that, but the changes will be reverted again
in future when we see that backend issue to be fixed.